+++ packages/connectanum_bench/example/parse_scenario.dart
@@ -1,0 +1,13 @@
+import 'package:connectanum_bench/connectanum_bench.dart';
+
+void main() {
+  final config = BenchmarkConfig.fromYaml('''
+benchmarks:
+  - name: websocket-smoke
+    type: wamp_rpc
+    duration: 5s
+    concurrency: 8
+''');
+
+  print(config.toPrettyJson());
+}